+++ src/fkooman/OAuth/Server/ApiService.php
@@ -27,9 +27,8 @@ use fkooman\Http\Exception\InternalServerErrorException;
 use fkooman\Http\Exception\NotFoundException;
 use fkooman\Http\Exception\ForbiddenException;
 use fkooman\Rest\Plugin\Bearer\BearerAuthentication;
-use fkooman\Rest\Plugin\Bearer\TokenIntrospection;
+use fkooman\Rest\Plugin\Bearer\TokenInfo;
 use fkooman\Rest\Plugin\Bearer\Scope;
-use fkooman\Rest\Plugin\Bearer\Entitlement;
 use InvalidArgumentException;
 
 class ApiService extends Service
@@ -40,10 +39,11 @@ class ApiService extends Service
     /** @var fkooman\OAuth\Server\IO */
     private $io;
 
-    public function __construct(PdoStorage $storage, IO $io = null)
+    public function __construct(PdoStorage $storage, Entitlements $entitlements, IO $io = null)
     {
         parent::__construct();
         $this->storage = $storage;
+        $this->entitlements = $entitlements;
 
         if (null === $io) {
             $io = new IO();
@@ -67,78 +67,78 @@ class ApiService extends Service
 
         $this->post(
             '/authorizations/',
-            function (Request $request, TokenIntrospection $tokenIntrospection) use ($compatThis) {
-                return $compatThis->postAuthorization($request, $tokenIntrospection);
+            function (Request $request, TokenInfo $tokenInfo) use ($compatThis) {
+                return $compatThis->postAuthorization($request, $tokenInfo);
             }
         );
 
         $this->get(
             '/authorizations/:id',
-            function (Request $request, TokenIntrospection $tokenIntrospection, $id) use ($compatThis) {
-                return $compatThis->getAuthorization($request, $tokenIntrospection, $id);
+            function (Request $request, TokenInfo $tokenInfo, $id) use ($compatThis) {
+                return $compatThis->getAuthorization($request, $tokenInfo, $id);
             }
         );
 
         $this->delete(
             '/authorizations/:id',
-            function (Request $request, TokenIntrospection $tokenIntrospection, $id) use ($compatThis) {
-                return $compatThis->deleteAuthorization($request, $tokenIntrospection, $id);
+            function (Request $request, TokenInfo $tokenInfo, $id) use ($compatThis) {
+                return $compatThis->deleteAuthorization($request, $tokenInfo, $id);
             }
         );
 
         $this->get(
             '/authorizations/',
-            function (Request $request, TokenIntrospection $tokenIntrospection) use ($compatThis) {
-                return $compatThis->getAuthorizations($request, $tokenIntrospection);
+            function (Request $request, TokenInfo $tokenInfo) use ($compatThis) {
+                return $compatThis->getAuthorizations($request, $tokenInfo);
             }
         );
 
         $this->get(
             '/applications/',
-            function (Request $request, TokenIntrospection $tokenIntrospection) use ($compatThis) {
-                return $compatThis->getApplications($request, $tokenIntrospection);
+            function (Request $request, TokenInfo $tokenInfo) use ($compatThis) {
+                return $compatThis->getApplications($request, $tokenInfo);
             }
         );
 
         $this->delete(
             '/applications/:id',
-            function (Request $request, TokenIntrospection $tokenIntrospection, $id) use ($compatThis) {
-                return $compatThis->deleteApplication($request, $tokenIntrospection, $id);
+            function (Request $request, TokenInfo $tokenInfo, $id) use ($compatThis) {
+                return $compatThis->deleteApplication($request, $tokenInfo, $id);
             }
         );
 
         $this->get(
             '/applications/:id',
-            function (Request $request, TokenIntrospection $tokenIntrospection, $id) use ($compatThis) {
-                return $compatThis->getApplication($request, $tokenIntrospection, $id);
+            function (Request $request, TokenInfo $tokenInfo, $id) use ($compatThis) {
+                return $compatThis->getApplication($request, $tokenInfo, $id);
             }
         );
 
         $this->post(
             '/applications/',
-            function (Request $request, TokenIntrospection $tokenIntrospection) use ($compatThis) {
-                return $compatThis->postApplication($request, $tokenIntrospection);
+            function (Request $request, TokenInfo $tokenInfo) use ($compatThis) {
+                return $compatThis->postApplication($request, $tokenInfo);
             }
         );
 
         $this->put(
             '/applications/:id',
-            function (Request $request, TokenIntrospection $tokenIntrospection, $id) use ($compatThis) {
-                return $compatThis->putApplication($request, $tokenIntrospection, $id);
+            function (Request $request, TokenInfo $tokenInfo, $id) use ($compatThis) {
+                return $compatThis->putApplication($request, $tokenInfo, $id);
             }
         );
 
         $this->get(
             '/stats/',
-            function (Request $request, TokenIntrospection $tokenIntrospection) use ($compatThis) {
-                return $compatThis->getStats($request, $tokenIntrospection);
+            function (Request $request, TokenInfo $tokenInfo) use ($compatThis) {
+                return $compatThis->getStats($request, $tokenInfo);
             }
         );
     }
 
-    public function postAuthorization(Request $request, TokenIntrospection $tokenIntrospection)
+    public function postAuthorization(Request $request, TokenInfo $tokenInfo)
     {
-        $this->requireScope($tokenIntrospection->getScope(), 'http://php-oauth.net/scope/authorize');
+        $this->requireScope($tokenInfo->getScope(), 'http://php-oauth.net/scope/authorize');
         $data = Json::decode($request->getContent());
         if (null === $data || !is_array($data) || !array_key_exists('client_id', $data) || !array_key_exists('scope', $data)) {
             throw new BadRequestException('missing client_id or scope');
@@ -154,8 +154,8 @@ class ApiService extends Service
         $refreshToken = (array_key_exists('refresh_token', $data) && $data['refresh_token']) ? $this->io->getRandomHex() : null;
 
         // check to see if an authorization for this client/resource_owner already exists
-        if (false === $this->storage->getApprovalByResourceOwnerId($clientId, $tokenIntrospection->getSub())) {
-            if (false === $this->storage->addApproval($clientId, $tokenIntrospection->getSub(), $data['scope'], $refreshToken)) {
+        if (false === $this->storage->getApprovalByResourceOwnerId($clientId, $tokenInfo->get('sub'))) {
+            if (false === $this->storage->addApproval($clientId, $tokenInfo->get('sub'), $data['scope'], $refreshToken)) {
                 throw new InternalServerErrorException('unable to add authorization');
             }
         } else {
@@ -170,10 +170,10 @@ class ApiService extends Service
         return $response;
     }
 
-    public function getAuthorization(Request $request, TokenIntrospection $tokenIntrospection, $id)
+    public function getAuthorization(Request $request, TokenInfo $tokenInfo, $id)
     {
-        $this->requireScope($tokenIntrospection->getScope(), 'http://php-oauth.net/scope/authorize');
-        $data = $this->storage->getApprovalByResourceOwnerId($id, $tokenIntrospection->getSub());
+        $this->requireScope($tokenInfo->getScope(), 'http://php-oauth.net/scope/authorize');
+        $data = $this->storage->getApprovalByResourceOwnerId($id, $tokenInfo->get('sub'));
         if (false === $data) {
             throw new NotFoundException('authorization not found');
         }
@@ -183,10 +183,10 @@ class ApiService extends Service
         return $response;
     }
 
-    public function deleteAuthorization(Request $request, TokenIntrospection $tokenIntrospection, $id)
+    public function deleteAuthorization(Request $request, TokenInfo $tokenInfo, $id)
     {
-        $this->requireScope($tokenIntrospection->getScope(), 'http://php-oauth.net/scope/authorize');
-        if (false === $this->storage->deleteApproval($id, $tokenIntrospection->getSub())) {
+        $this->requireScope($tokenInfo->getScope(), 'http://php-oauth.net/scope/authorize');
+        if (false === $this->storage->deleteApproval($id, $tokenInfo->get('sub'))) {
             throw new NotFoundException('authorization not found');
         }
         $response = new JsonResponse(200);
@@ -195,10 +195,10 @@ class ApiService extends Service
         return $response;
     }
 
-    public function getAuthorizations(Request $request, TokenIntrospection $tokenIntrospection)
+    public function getAuthorizations(Request $request, TokenInfo $tokenInfo)
     {
-        $this->requireScope($tokenIntrospection->getScope(), 'http://php-oauth.net/scope/authorize');
-        $data = $this->storage->getApprovals($tokenIntrospection->getSub());
+        $this->requireScope($tokenInfo->getScope(), 'http://php-oauth.net/scope/authorize');
+        $data = $this->storage->getApprovals($tokenInfo->get('sub'));
 
         $response = new JsonResponse(200);
         $response->setContent($data);
@@ -206,11 +206,11 @@ class ApiService extends Service
         return $response;
     }
 
-    public function getApplications(Request $request, TokenIntrospection $tokenIntrospection)
+    public function getApplications(Request $request, TokenInfo $tokenInfo)
     {
-        $this->requireScope($tokenIntrospection->getScope(), 'http://php-oauth.net/scope/manage');
-        $this->requireEntitlement($tokenIntrospection->getEntitlement(), 'http://php-oauth.net/entitlement/manage');
-        // do not require entitlement to list clients...
+        $this->requireScope($tokenInfo->getScope(), 'http://php-oauth.net/scope/manage');
+        $this->requireEntitlement($tokenInfo->get('sub'), 'http://php-oauth.net/entitlement/manage');
+
         $data = $this->storage->getClients();
         $response = new JsonResponse(200);
         $response->setContent($data);
@@ -218,10 +218,11 @@ class ApiService extends Service
         return $response;
     }
 
-    public function deleteApplication(Request $request, TokenIntrospection $tokenIntrospection, $id)
+    public function deleteApplication(Request $request, TokenInfo $tokenInfo, $id)
     {
-        $this->requireScope($tokenIntrospection->getScope(), 'http://php-oauth.net/scope/manage');
-        $this->requireEntitlement($tokenIntrospection->getEntitlement(), 'http://php-oauth.net/entitlement/manage');
+        $this->requireScope($tokenInfo->getScope(), 'http://php-oauth.net/scope/manage');
+        $this->requireEntitlement($tokenInfo->get('sub'), 'http://php-oauth.net/entitlement/manage');
+
         if (false === $this->storage->deleteClient($id)) {
             throw new NotFoundException('application not found');
         }
@@ -231,10 +232,11 @@ class ApiService extends Service
         return $response;
     }
 
-    public function getApplication(Request $request, TokenIntrospection $tokenIntrospection, $id)
+    public function getApplication(Request $request, TokenInfo $tokenInfo, $id)
     {
-        $this->requireScope($tokenIntrospection->getScope(), 'http://php-oauth.net/scope/manage');
-        $this->requireEntitlement($tokenIntrospection->getEntitlement(), 'http://php-oauth.net/entitlement/manage');
+        $this->requireScope($tokenInfo->getScope(), 'http://php-oauth.net/scope/manage');
+        $this->requireEntitlement($tokenInfo->get('sub'), 'http://php-oauth.net/entitlement/manage');
+
         $data = $this->storage->getClient($id);
         if (false === $data) {
             throw new NotFoundException('application not found');
@@ -245,10 +247,10 @@ class ApiService extends Service
         return $response;
     }
 
-    public function postApplication(Request $request, TokenIntrospection $tokenIntrospection)
+    public function postApplication(Request $request, TokenInfo $tokenInfo)
     {
-        $this->requireScope($tokenIntrospection->getScope(), 'http://php-oauth.net/scope/manage');
-        $this->requireEntitlement($tokenIntrospection->getEntitlement(), 'http://php-oauth.net/entitlement/manage');
+        $this->requireScope($tokenInfo->getScope(), 'http://php-oauth.net/scope/manage');
+        $this->requireEntitlement($tokenInfo->get('sub'), 'http://php-oauth.net/entitlement/manage');
 
         $clientData = null;
         try {
@@ -271,10 +273,10 @@ class ApiService extends Service
         return $response;
     }
 
-    public function putApplication(Request $request, TokenIntrospection $tokenIntrospection, $id)
+    public function putApplication(Request $request, TokenInfo $tokenInfo, $id)
     {
-        $this->requireScope($tokenIntrospection->getScope(), 'http://php-oauth.net/scope/manage');
-        $this->requireEntitlement($tokenIntrospection->getEntitlement(), 'http://php-oauth.net/entitlement/manage');
+        $this->requireScope($tokenInfo->getScope(), 'http://php-oauth.net/scope/manage');
+        $this->requireEntitlement($tokenInfo->get('sub'), 'http://php-oauth.net/entitlement/manage');
 
         $clientData = null;
         try {
@@ -295,10 +297,11 @@ class ApiService extends Service
         return $response;
     }
 
-    public function getStats(Request $request, TokenIntrospection $tokenIntrospection)
+    public function getStats(Request $request, TokenInfo $tokenInfo)
     {
-        $this->requireScope($tokenIntrospection->getScope(), 'http://php-oauth.net/scope/manage');
-        $this->requireEntitlement($tokenIntrospection->getEntitlement(), 'http://php-oauth.net/entitlement/manage');
+        $this->requireScope($tokenInfo->getScope(), 'http://php-oauth.net/scope/manage');
+        $this->requireEntitlement($tokenInfo->get('sub'), 'http://php-oauth.net/entitlement/manage');
+
         $data = $this->storage->getStats();
 
         $response = new JsonResponse(200);
@@ -314,9 +317,10 @@ class ApiService extends Service
         }
     }
 
-    private function requireEntitlement(Entitlement $entitlement, $entitlementValue)
+    private function requireEntitlement($userId, $entitlementValue)
     {
-        if (!$entitlement->hasEntitlement($entitlementValue)) {
+        $entitlements = $this->entitlements->getEntitlement($userId);
+        if (array_key_exists($entitlementValue, $entitlements)) {
             throw new ForbiddenException('insufficient_entitlement', sprintf('need entitlement "%s"', $entitlementValue));
         }
     }