+++ web/introspect.php
@@ -22,7 +22,6 @@ use fkooman\OAuth\Server\TokenIntrospectionService;
 use fkooman\OAuth\Server\PdoStorage;
 use fkooman\Http\Exception\HttpException;
 use fkooman\Http\Exception\InternalServerErrorException;
-use fkooman\OAuth\Server\Entitlements;
 
 set_error_handler(
     function ($errno, $errstr, $errfile, $errline) {
@@ -42,11 +41,9 @@ try {
     );
 
     $tokenIntrospectionService = new TokenIntrospectionService(
-        new PdoStorage($db),
-        new Entitlements($iniReader->v('entitlementsFile'))
+        new PdoStorage($db)
     );
     $tokenIntrospectionService->run()->sendResponse();
 } catch (Exception $e) {
-    error_log($e->getMessage());
     TokenIntrospectionService::handleException($e)->sendResponse();
 }